# openapi.yaml
openapi: 3.1.0
info:
  title: Synthetic Data Generation Service
  version: 1.0.0
  description: |
    REST API для сервиса генерации конфиденциальных синтетических табличных данных
    с применением дифференциальной приватности (DP-CTGAN).

    ## Основной сценарий использования

    1. Загрузить датасет — `POST /datasets`
    2. Загрузить конфигурацию — `POST /configs`
    3. Запустить пайплайн — `POST /runs`
    4. Опрашивать статус — `GET /runs/{run_id}` (или дождаться webhook)
    5. Скачать синтетику — `GET /runs/{run_id}/synthetic`
    6. Изучить отчёт — `GET /runs/{run_id}/report`

    ## Асинхронное выполнение

    Обучение DP-CTGAN занимает несколько минут. Все запуски пайплайна
    (`POST /runs`) возвращают `run_id` немедленно со статусом `queued`.
    Итоговый статус — `completed` | `failed` | `cancelled`.

    ## Авторизация

    Все запросы требуют заголовок `Authorization: Bearer <token>`.
  contact:
    name: Сергей Рогачев
    email: s.rogachev@example.com
  license:
    name: MIT
servers:
- url: http://localhost:8000/api/v1
  description: Локальная разработка
- url: https://synth.internal/api/v1
  description: Внутренний production-стенд
...
security:
- BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          examples:
          - NOT_FOUND
        message:
          type: string
          examples:
          - Запуск с указанным ID не найден
        details:
          type: object
          additionalProperties: true
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          examples:
          - 42
        page:
          type: integer
          examples:
          - 1
        per_page:
          type: integer
          examples:
          - 20
        pages:
          type: integer
          examples:
          - 3
...
    ConfigSummary:
      type: object
      properties:
        name:
          type: string
          examples:
          - adult_epsilon3
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        dataset_name:
          type:
          - string
          - 'null'
          description: Датасет, для которого предназначен конфиг
        epsilon:
          type: number
          examples:
          - 3.0
        epochs:
          type: integer
          examples:
          - 300
    ConfigValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                examples:
                - generator.epsilon
              message:
                type: string
                examples:
                - epsilon должен быть > 0
          description: Пустой массив если valid=true
        warnings:
          type: array
          items:
            type: string
          description: Некритичные предупреждения (например, preprocessor_eps близко к epsilon)
          examples:
          - - preprocessor_eps занимает >30% от epsilon — останется мало бюджета на обучение
    DatasetSummary:
      type: object
      properties:
        name:
          type: string
          examples:
          - adult_census
        uploaded_at:
          type: string
          format: date-time
        rows:
          type: integer
          examples:
          - 48842
        columns:
          type: integer
          examples:
          - 15
        file_size_bytes:
          type: integer
          examples:
          - 3981312
    DatasetSchema:
      type: object
      properties:
        categorical:
          type: array
          items:
            type: string
          examples:
          - - workclass
            - education
            - occupation
            - income
        continuous:
          type: array
          items:
            type: string
          examples:
          - - age
            - hours-per-week
            - capital-gain
        ignored:
          type: array
          items:
            type: string
          examples:
          - - fnlwgt
        detected_at:
          type: string
          format: date-time
    DatasetPreview:
      type: object
      properties:
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        stats:
          type: object
          description: Базовая описательная статистика по каждой колонке
          additionalProperties:
            type: object
            properties:
              dtype:
                type: string
              null_count:
                type: integer
              unique_count:
                type: integer
              mean:
                type:
                - number
                - 'null'
              min:
                type:
                - string
                - 'null'
              max:
                type:
                - string
                - 'null'
...
  parameters:
...
    ConfigName:
      name: name
      in: path
      required: true
      schema:
        type: string
      description: Имя конфигурации (без расширения .yaml)
      examples:
        default:
          value: adult_epsilon3
    DatasetName:
      name: name
      in: path
      required: true
      schema:
        type: string
      description: Имя датасета
      examples:
        default:
          value: adult_census
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  responses:
    BadRequest:
      description: Невалидный запрос (ошибка валидации входных данных)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: VALIDATION_ERROR
                message: n_rows должен быть > 0
    Unauthorized:
      description: Отсутствует или невалиден токен авторизации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: UNAUTHORIZED
                message: Bearer token обязателен
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                code: NOT_FOUND
                message: Запуск не найден
...
paths:
...
  /configs:
    get:
      tags:
      - configs
      summary: Список конфигураций
      operationId: listConfigs
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConfigSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - configs
      summary: Загрузить конфигурацию
      description: Принимает YAML-файл, валидирует его и сохраняет под указанным именем.
      operationId: createConfig
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - name
              - file
              properties:
                name:
                  type: string
                  description: Имя конфига (без .yaml)
                  examples:
                  - adult_epsilon3
                file:
                  type: string
                  format: binary
                  description: YAML-файл конфигурации
      responses:
        '201':
          description: Конфиг создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigSummary'
        '400':
          description: Невалидный YAML или ошибка валидации конфига
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigValidationResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Конфиг с таким именем уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /configs/validate:
    post:
      tags:
      - configs
      summary: Валидировать конфигурацию
      description: Проверяет YAML на корректность без сохранения. Аналог `cli.py --check`.
      operationId: validateConfig
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
          application/x-yaml:
            schema:
              type: string
      responses:
        '200':
          description: Результат валидации (valid может быть false — это не ошибка HTTP)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigValidationResult'
              examples:
                valid:
                  summary: Конфиг валиден
                  value:
                    valid: true
                    errors: []
                    warnings:
                    - preprocessor_eps занимает >30% от epsilon
                invalid:
                  summary: Конфиг невалиден
                  value:
                    valid: false
                    errors:
                    - field: generator.epsilon
                      message: epsilon должен быть > 0
                    warnings: []
        '401':
          $ref: '#/components/responses/Unauthorized'
  /configs/{name}:
    get:
      tags:
      - configs
      summary: Получить конфигурацию
      operationId: getConfig
      parameters:
      - $ref: '#/components/parameters/ConfigName'
      responses:
        '200':
          description: OK
          content:
            application/x-yaml:
              schema:
                type: string
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
      - configs
      summary: Заменить конфигурацию
      description: Полностью заменяет существующий конфиг. Валидирует перед сохранением.
      operationId: replaceConfig
      parameters:
      - $ref: '#/components/parameters/ConfigName'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Конфиг обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigSummary'
        '400':
          description: Ошибка валидации нового конфига
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigValidationResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - configs
      summary: Удалить конфигурацию
      operationId: deleteConfig
      parameters:
      - $ref: '#/components/parameters/ConfigName'
      responses:
        '204':
          description: Конфиг удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Конфиг используется в активном запуске
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                default:
                  value:
                    code: CONFIG_IN_USE
                    message: Конфиг используется в запуске 3fa85f64-...
  /datasets:
    get:
      tags:
      - datasets
      summary: Список датасетов
      operationId: listDatasets
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DatasetSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - datasets
      summary: Загрузить датасет
      description: |
        Принимает CSV-файл, вычисляет базовую статистику и сохраняет.
        Автодетекция схемы колонок запускается асинхронно после загрузки.
      operationId: uploadDataset
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - name
              - file
              properties:
                name:
                  type: string
                  description: Имя датасета (уникальное)
                  examples:
                  - adult_census
                file:
                  type: string
                  format: binary
                  description: CSV-файл
                na_values:
                  type: string
                  description: Разделённые запятой значения, трактуемые как NaN
                  examples:
                  - ?,NA,N/A
      responses:
        '201':
          description: Датасет загружен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Датасет с таким именем уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /datasets/{name}:
    delete:
      tags:
      - datasets
      summary: Удалить датасет
      operationId: deleteDataset
      parameters:
      - $ref: '#/components/parameters/DatasetName'
      responses:
        '204':
          description: Датасет удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Датасет используется в активном запуске
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /datasets/{name}/schema:
    get:
      tags:
      - datasets
      summary: Автодетектированная схема колонок
      description: |
        Возвращает результат `DataProcessor.detect_column_types()`:
        разбивку колонок на categorical / continuous / ignored.
        Используйте как основу для составления конфига.
      operationId: getDatasetSchema
      parameters:
      - $ref: '#/components/parameters/DatasetName'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetSchema'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /datasets/{name}/preview:
    post:
      tags:
      - datasets
      summary: Предпросмотр и статистика датасета
      operationId: previewDataset
      parameters:
      - $ref: '#/components/parameters/DatasetName'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                n_rows:
                  type: integer
                  default: 10
                  minimum: 1
                  maximum: 1000
                  description: Количество строк для предпросмотра
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetPreview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...